pyproject.toml
[project]
name = "apple-health-data"
version = "0.1.0"
description = "Extract Hevy workout data to S3 and transform with dbt"
readme = "README.md"
requires-python = "==3.12.0"
dependencies = [
    "dlt[filesystem,parquet]==1.21.0",
    "s3fs==2026.1.0",
    "pyarrow==23.0.0",
    "polars==1.38.0",
    "dbt-core==1.11.2",
    "dbt-duckdb==1.10.0",
    "python-dotenv==1.2.1",
    "streamlit==1.54.0",
    "altair==6.0.0",
    "beautifulsoup4>=4.14.3",
    "requests>=2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pipelines", "src/dashboard"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
# Streamlit pages require st.set_page_config() before other imports
"src/dashboard/pages/*.py" = ["E402"]
"src/dashboard/Home.py" = ["E402"]

# =============================================================================
# Dashboard Configuration (non-sensitive values)
# =============================================================================
[tool.dashboard]
# S3 bucket and prefix for transformed data
s3_bucket_name = "api-health-data-ntonthat"
s3_transformed_prefix = "transformed"
aws_region = "ap-southeast-2"

# User info
user_name = "Nam Tonthat"
openpowerlifting_url = "https://www.openpowerlifting.org/u/namtonthat"

[tool.dashboard.goals]
sleep_hours = 7.0
sleep_deep_hours = 1.5
sleep_rem_hours = 1.5
sleep_light_hours = 3.5
protein_g = 170.0
carbs_g = 300.0
fat_g = 60.0
steps = 10000

[dependency-groups]
dev = [
    "pytest==9.0.2",
    "ruff==0.15.0",
    "sqlfluff==3.3.1",
    "sqlfluff-templater-dbt==3.3.1",
    "pre-commit>=4.0",
]
